refactor: move payment request table data
