+++ packages/cryptoplease/lib/features/pending_activities/pending_activities_repository.dart
@@ -1,6 +1,7 @@
 import 'package:cryptoplease/data/db/db.dart';
 import 'package:cryptoplease/features/outgoing_direct_payments/bl/repository.dart';
 import 'package:cryptoplease/features/outgoing_split_key_payments/bl/repository.dart';
+import 'package:cryptoplease/features/payment_request/bl/repository.dart';
 import 'package:cryptoplease/features/pending_activities/pending_activity.dart';
 import 'package:drift/drift.dart';
 import 'package:fast_immutable_collections/fast_immutable_collections.dart';
